+++ config/autoload/dependencies.global.php
@@ -13,7 +13,7 @@ return [
         'factories' => [
             Application::class => ApplicationFactory::class,
             Helper\UrlHelper::class => Helper\UrlHelperFactory::class,
-            App\Service\MeetupsServiceInterface::class => App\Service\MeetupsServiceFactory::class,
+            App\Service\Meetup\MeetupsServiceInterface::class => App\Service\Meetup\MeetupsServiceFactory::class,
             Doctrine\ORM\EntityManagerInterface::class => ContainerInteropDoctrine\EntityManagerFactory::class,
         ],
         'delegators' => [
