+++ src/App/Service/Meetup/MeetupsServiceInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace App\Service;
+namespace App\Service\Meetup;
 
 use App\Entity\Meetup;
 use DateTimeImmutable;
